# Request-level bail