--- a/src/chain/context-handler-impl.ts
+++ b/src/chain/context-handler-impl.ts
@@ -3,13 +3,13 @@ import { Optional } from '../context';
 import { ChainCondition, CustomCondition } from '../context-items';
 import { CustomValidator } from '../base';
 import { Bail } from '../context-items/bail';
-import { ContextHandler } from './context-handler';
+import { BailOptions, ContextHandler } from './context-handler';
 import { ContextRunner } from './context-runner';
 
 export class ContextHandlerImpl<Chain> implements ContextHandler<Chain> {
   constructor(private readonly builder: ContextBuilder, private readonly chain: Chain) {}
 
-  bail(opts?: { level?: 'chain' | 'request' }) {
+  bail(opts?: BailOptions) {
     if (opts?.level === 'request') {
       this.builder.setRequestBail();
     }