--- a/src/chain/context-handler.ts
+++ b/src/chain/context-handler.ts
@@ -23,7 +23,18 @@ export interface ContextHandler<Chain> {
    *
    * @returns the current validation chain
    */
-  bail(): Chain;
+  bail(opts?: {
+    /**
+     * Defines the level at which to stop running further validations:
+     * - When set to `chain`, further validations won't be run for this validation chain if there
+     *   are any errors.
+     * - When set to `request`, no further validations on the same request will be run either if
+     *   there are any errors.
+     *
+     * @default 'chain'
+     */
+    level?: 'chain' | 'request';
+  }): Chain;
 
   /**
    * Adds a condition on whether the validation should continue on a field or not.