--- a/src/middlewares/exact.ts
+++ b/src/middlewares/exact.ts
@@ -105,7 +105,7 @@ export function checkExact(
       unknownFields = unknownFields.concat(selectUnknownFields(req, fields, [location]));
     }
 
-    const context = new Context([], [], [], false);
+    const context = new Context([], [], [], false, false);
     if (unknownFields.length) {
       context.addError({
         type: 'unknown_fields',