--- a/src/utils.ts
+++ b/src/utils.ts
@@ -1,3 +1,6 @@
+import { Request } from './base';
+import { ContextRunningOptions, ResultWithContext, ValidationChainLike } from './chain';
+
 export const bindAll = <T>(object: T): { [K in keyof T]: T[K] } => {
   const protoKeys = Object.getOwnPropertyNames(Object.getPrototypeOf(object)) as (keyof T)[];
   protoKeys.forEach(key => {
@@ -23,4 +26,35 @@ export function toString(value: any): string {
   }
 
   return String(value);
+}
+
+/**
+ * Runs all validation chains, and returns their results.
+ *
+ * If one of them has a request-level bail set, the previous chains will be awaited on so that
+ * results are not skewed, which can be slow.
+ * If this same chain also contains errors, no further chains are run.
+ */
+export async function runAllChains(
+  req: Request,
+  chains: readonly ValidationChainLike[],
+  runOpts?: ContextRunningOptions,
+): Promise<ResultWithContext[]> {
+  const promises: Promise<ResultWithContext>[] = [];
+  for (const chain of chains) {
+    const bails = chain.builder.build().bail;
+    if (bails) {
+      await Promise.all(promises);
+    }
+
+    const resultPromise = chain.run(req, runOpts);
+    promises.push(resultPromise);
+    if (bails) {
+      const result = await resultPromise;
+      if (!result.isEmpty()) {
+        break;
+      }
+    }
+  }
+  return Promise.all(promises);
 }