## Description

Closes #1100

Implements an option to `.bail()` so that no further contexts will run if there are any errors.

Supports:
- [x] `oneOf`
- [x] `checkSchema`
- [x] `checkExact`


⚠️ This is a perf-loss for those using any of the above functions, as the chains now need to be awaited on, otherwise results will be unreliable.

Suppose that you have a request with 1k fields, one validator on each field, each taking 10ms to complete.
Also suppose that no fields present errors, so every validator runs.

On my machine, the time it takes for running all validation chains looks like this:
- **With request-level bail on every field:** 10s
- **With request-level bail on half of all fields:** 10s (no difference)
- **With request-level bail on a third of all fields:** ~7s (roughly 1/3 better than the above scenarios)
- **Without request-level bail:** ~100 ms.

## To-do list

<!-- Put an "x" to indicate you've done each of the following -->

- [x] I have added tests for what I changed.

<!-- If this pull request isn't ready, add any remaining tasks here -->

- [x] This pull request is ready to merge.